Safari-611.1.10/LayoutTests/http/tests/plugins/local-geturl-from-remote.html

<html>
<head>
<script>
function runTest()
{
    if (window.testRunner)
        testRunner.dumpAsText();
        
    // No need to let this point to a real file.

    
    var result = plg.getURL('file:///tmp.html', '_self');
    if (result == 1)
        document.getElementById('result').innerHTML = "SUCCESS";
    else
        document.getElementById('result').innerHTML = "FAILURE";
}
</script>
</head>
<body onload="runTest()">
<div>This tests that a plug-in in a remote document can't access local files using NPN_GetURL</div>
<embed name="plg" type="application/x-webkit-test-netscape"></embed>
<div id="result">FAILURE</div>
</body>
</html>
